web/requirements.txt
Flask==1.0.2
Flask-SocketIO==3.1.2
redis
pyserial
pyax12
